FIX list when browsing issues
I had two issues with this plugin :

1. It won't work on plain /repos/issues url, because it expects '?' parameters
2. It only work on initial page load, not after browsing and getting
   content replaced with ajax

So, first, for the easy part, I changed location regexp and handled
undefined parameter match to let initialize with '?' query string.

Then, I tried to make it load after page change. I first tried to use
onpopstate event, but for some reason, it won't work when using
pushstate (I suppose github bound it own callback and prevented
propagation).

So I decided to reload the assignee list on $.ajaxComplete.

Problem was, the extension has its own version of jquery, which get
overwritten by github one, so callback are simply discards.

But since extension is ran before in page jquery is loaded, we could not
simply remove jquery from extension and use the in page one.

So I removed filter_by_assignee.js from js files loaded by extension,
and added a script which inject it after page load, so it runs when in
page jquery is already there.

Thanks for this extension, btw, it's a feature that github miss a lot
when I tried to explain that project management using github is way
enough :)